de integer a float
git-svn-id: https://subversion.assembla.com/svn/uadepfi2011@23 1215cf96-3890-4680-975f-d2b89ea09188

+++ core/src/main/java/com/uade/pfi/core/TransportLocation.java
@@ -8,7 +8,7 @@ public class TransportLocation {
 	public TransportLocation() {
 	}
 	
-	public TransportLocation(Integer latitude, Integer longitude, String name) {
+	public TransportLocation(Float latitude, Float longitude, String name) {
 		this.location = new Location(latitude,longitude);
 		this.name = name;
 	}
